+++ script/check-s3-images.js
@@ -9,7 +9,7 @@ const allVersions = require('../lib/all-versions')
 const nonEnterpriseDefaultVersion = require('../lib/non-enterprise-default-version')
 const { getS3BucketPathFromVersion, getVersionFromS3BucketPath } = require('../lib/s3-bucket-path-utils')
 const patterns = require('../lib/patterns')
-const authenticateToAWS = require('../lib/authenticate-to-aws.js')
+const authenticateToAWS = require('./helpers/authenticate-to-aws.js')
 const readlineSync = require('readline-sync')
 const { execSync } = require('child_process')
 const enterpriseServerVersions = Object.keys(allVersions).filter(v => v.startsWith('enterprise-server@'))
