+++ script/graphql/update-files.js
@@ -7,7 +7,7 @@ const yaml = require('js-yaml')
 const { execSync } = require('child_process')
 const graphqlDataDir = path.join(process.cwd(), 'data/graphql')
 const graphqlStaticDir = path.join(process.cwd(), 'lib/graphql/static')
-const { getContents, listMatchingRefs } = require('../../lib/git-utils')
+const { getContents, listMatchingRefs } = require('../helpers/git-utils')
 const dataFilenames = require('./utils/data-filenames')
 const allVersions = require('../../lib/all-versions')
 const processPreviews = require('./utils/process-previews')
