+++ script/helpers/authenticate-to-aws.js
