+++ script/update-enterprise-dates.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const { getContents } = require('../lib/git-utils')
+const { getContents } = require('./helpers/git-utils')
 const fs = require('fs')
 const path = require('path')
 const enterpriseDatesFile = path.join(__dirname, '../lib/enterprise-dates.json')
