+++ script/upload-images-to-s3.js
@@ -4,7 +4,7 @@ const fs = require('fs')
 const path = require('path')
 const program = require('commander')
 const allVersions = require('../lib/all-versions')
-const authenticateToAWS = require('../lib/authenticate-to-aws')
+const authenticateToAWS = require('./helpers/authenticate-to-aws')
 const nonEnterpriseDefaultVersion = require('../lib/non-enterprise-default-version')
 const { getS3BucketPathFromVersion } = require('../lib/s3-bucket-path-utils')
 const walk = require('walk-sync')
