+++ tests/content/crowdin-config.js
@@ -1,4 +1,4 @@
-const config = require('../../lib/crowdin-config').read()
+const config = require('../helpers/crowdin-config').read()
 const { loadPages } = require('../../lib/pages')
 const ignoredPagePaths = config.files[0].ignore
 const ignoredDataPaths = config.files[2].ignore
